requirements.txt
bcrypt==4.0.0
click==8.1.3
cpf-generator==2.1.4
deprecation==2.1.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.3
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
packaging==21.3
protobuf==3.20.1
pyparsing==3.0.9
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
